feat(controllers): :sparkles: add check for non-empty attributes before deleting and updating product attributes

+++ app/Http/Controllers/Admin/ProductController.php
@@ -154,22 +154,23 @@ class ProductController extends Controller
 
                 $product_id = $product->id;
 
-                // Delete all attributes for the product and insert new ones
-                ProductAttribute::where('product_id', $product_id)->delete();
+                if ($request->attribute != '') {
+                    ProductAttribute::where('product_id', $product_id)->delete();
 
-                foreach ($request->attribute as $key => $value) {
-                    ProductAttribute::updateOrCreate(
-                        [
-                            'product_id' => $product_id,
-                            'category_id' => $request->category,
-                            'attribute_value_id' => $value,
-                        ],
-                        [
-                            'product_id' => $product_id,
-                            'category_id' => $request->category,
-                            'attribute_value_id' => $value,
-                        ]
-                    );
+                    foreach ($request->attribute as $key => $value) {
+                        ProductAttribute::updateOrCreate(
+                            [
+                                'product_id' => $product_id,
+                                'category_id' => $request->category,
+                                'attribute_value_id' => $value,
+                            ],
+                            [
+                                'product_id' => $product_id,
+                                'category_id' => $request->category,
+                                'attribute_value_id' => $value,
+                            ]
+                        );
+                    }
                 }
 
                 foreach ($request->sku as $key => $value) {
